ShapeManager.cs splited in two files
