+++ ShapeFilesParser.Business/ShapeManager.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ShapeFilesParser.Business.Parsers;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -7,44 +8,6 @@ using System.Threading.Tasks;
 
 namespace ShapeFilesParser.Business
 {
-    public abstract class BaseShapeParser<T>
-    {
-        public abstract T Parse(byte[] recordContent);
-
-        public int ReadShort(byte[] source, int index, bool littleEndian)
-        {
-            var data = new byte[2];
-            for (int i = 0; i < 2; i++)
-            {
-                var currentIndex = index + (littleEndian ? 1 - i : i);
-                data[i] = source[currentIndex];
-            }
-            return (data[0] << 8) | (data[1]);
-        }
-
-        public int ReadInt(byte[] source, int index, bool littleEndian)
-        {
-            var data = new byte[4];
-            for (int i = 0; i < 4; i++)
-            {
-                var currentIndex = index + (littleEndian ? 3 - i : i);
-                data[i] = source[currentIndex];
-            }
-            return (data[0] << 24) | (data[1] << 16) | (data[2] << 8) | (data[3]);
-        }
-
-        public double ReadDouble(byte[] source, int index, bool littleEndian)
-        {
-            var data = new byte[8];
-            for (int i = 0; i < 8; i++)
-            {
-                var currentIndex = index + (littleEndian != BitConverter.IsLittleEndian ? 7 - i : i);
-                data[i] = source[currentIndex];
-            }
-            return BitConverter.ToDouble(data, 0);
-        }
-    }
-
     public class ShapeManager<T> where T : new()
     {
         private BaseShapeParser<T> parser;
